+++ tests/SelfTest/TestRegistrations.cpp
@@ -4,6 +4,7 @@
  */
 
 #include <catch2/catch_tag_alias_autoregistrar.hpp>
+#include <catch2/reporters/catch_reporter_event_listener.hpp>
 
 // Some example tag aliases
 CATCH_REGISTER_TAG_ALIAS( "[@nhf]", "[failing]~[.]" )
@@ -15,10 +16,9 @@ CATCH_REGISTER_TAG_ALIAS( "[@tricky]", "[tricky]~[.]" )
 #   pragma clang diagnostic ignored "-Wc++98-compat"
 #endif
 
-#include <catch2/reporters/catch_reporter_streaming_base.hpp>
 
-struct TestListener : Catch::TestEventListenerBase {
-    using TestEventListenerBase::TestEventListenerBase;
+struct TestListener : Catch::EventListenerBase {
+    using EventListenerBase::EventListenerBase;
 };
 
 #include <catch2/catch_reporter_registrars.hpp>